rename data -> modify, and _GrpcCollection -> _QueryCollection
