+++ weaviate/collection/collection.py
@@ -21,8 +21,8 @@ from weaviate.collection.classes.grpc import MetadataQuery, PROPERTIES
 from weaviate.collection.classes.types import Properties, TProperties, _check_data_model
 from weaviate.collection.collection_base import CollectionBase, CollectionObjectBase
 from weaviate.collection.config import _ConfigCollection
-from weaviate.collection.data import _DataCollection
-from weaviate.collection.grpc import _GrpcCollection
+from weaviate.collection.modify import _ModifyCollection
+from weaviate.collection.query import _QueryCollection
 from weaviate.collection.object_iterator import _ObjectIterator
 from weaviate.collection.tenants import _Tenants
 from weaviate.connect import Connection
@@ -43,9 +43,11 @@ class CollectionObject(CollectionObjectBase, Generic[TProperties]):
         self._connection = connection
 
         self.config = _ConfigCollection(self._connection, name)
-        self.data = _DataCollection[TProperties](connection, name, consistency_level, tenant, type_)
-        self.query = _GrpcCollection[TProperties](
-            connection, name, self.data, consistency_level, tenant
+        self.modify = _ModifyCollection[TProperties](
+            connection, name, consistency_level, tenant, type_
+        )
+        self.query = _QueryCollection[TProperties](
+            connection, name, self.modify, consistency_level, tenant
         )
         self.tenants = _Tenants(connection, name)
