+++ weaviate/collection/collection_model.py
@@ -12,8 +12,8 @@ from weaviate.collection.classes.orm import (
 )
 from weaviate.collection.collection_base import CollectionBase, CollectionObjectBase
 from weaviate.collection.config import _ConfigCollectionModel
-from weaviate.collection.data import _DataCollectionModel
-from weaviate.collection.grpc import _GrpcCollectionModel
+from weaviate.collection.modify import _DataCollectionModel
+from weaviate.collection.query import _GrpcCollectionModel
 from weaviate.collection.tenants import _Tenants
 from weaviate.connect import Connection
 from weaviate.exceptions import UnexpectedStatusCodeException
